+++ commons/app-commons/src/main/java/com/becker/freelance/commons/ExecutionConfiguration.java
@@ -4,8 +4,9 @@ import com.becker.freelance.commons.pair.Pair;
 import com.becker.freelance.commons.timeseries.TimeSeries;
 
 import java.time.LocalDateTime;
+import com.becker.freelance.math.Decimal;
 
-public record ExecutionConfiguration(Pair pair, double initialWalletAmount, TimeSeries eurUsd, LocalDateTime startTime,
+public record ExecutionConfiguration(Pair pair, Decimal initialWalletAmount, TimeSeries eurUsd, LocalDateTime startTime,
                                      LocalDateTime endTime) {
 
     public TimeSeries getEurUsdTimeSeries() {
